Fix noise choppiness and audio_io type errors
- Revert is_unmodulated=true change that caused choppy audio
- Keep original logic: static calibration only for unmodulated noise
- Increase OLA hysteresis to 0.20 (20%) for more stability
- Use moderate OLA smoothing coefficient 0.9995 (was too aggressive at 0.998)
- Fix sample_rate type mismatch in audio_io.rs (f32 -> u32 cast)

+++ native/realtime_backend/src/audio_io.rs
@@ -254,7 +254,7 @@ pub fn run_audio_stream<C>(
     }
 
     let channels = 2usize;
-    let sample_rate = scheduler.sample_rate;
+    let sample_rate = scheduler.sample_rate as u32;
     let max_samples = samples_for_seconds(sample_rate, AUDIO_RING_MAX_SECONDS, channels)
         .max(AUDIO_WORKER_BLOCK_FRAMES * channels);
     let rb = HeapRb::<f32>::new(max_samples);
@@ -366,7 +366,7 @@ fn run_audio_stream_android<C>(
     log::error!("REALTIME_BACKEND: Starting Oboe stream (Android specialized)...");
 
     let channels = 2usize;
-    let sample_rate = scheduler.sample_rate;
+    let sample_rate = scheduler.sample_rate as u32;
     let max_samples = samples_for_seconds(sample_rate, AUDIO_RING_MAX_SECONDS, channels)
         .max(AUDIO_WORKER_BLOCK_FRAMES * channels);
     let rb = HeapRb::<f32>::new(max_samples);
